Supports FID scan for FID or spectral shimming
Acquisition does not support wbs processing so switched to
wnt and wexp. To abort FID scan, switched from aa to wexp=''
This keeps data in the experiment so it can be used for interactive
adjustment of vs, pan and zoom, etc. Need to allow enough time
between wnt processing (wft ds) and subsequent wexp processing (au),
which removes the data from the exeriment, so that the FID or spectrum
can be displayed.