FID scan would stop prematurely. Infoproc could fail to start.
There was a race between starting a new B12proc due to an
au call from wexp processing and the exit of the B12proc
that initiated the wexp processing. This resulted in the new
B12proc failing to start and the FID scan stopping. Now, if
the previous B12proc has not yet exited by the time the new B12proc
is requested, a new B12proc will be started upon receiving
the SIGCHLD from the exiting B12proc.

Infoproc would exit if it failed to set up the rpc server system.
This is only used for distributed ft3d processing and the obsolete
remote Acqstat display. Neither of these are an issue for B12
so the exit was removed.